created two different screens for todo and completed
