+++ lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:todo_taxfyle/screens/todo_screen.dart';
 
 void main() => runApp(MyApp());
 
@@ -6,38 +7,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      home: DefaultTabController(
-        length: 2,
-        child: Scaffold(
-          appBar: AppBar(
-            title: Text('Todo App'),
-            bottom: TabBar(
-              tabs: <Widget>[
-                Text('TODO'),
-                Text('COMPLETED'),
-              ],
-            ),
-          ),
-          body: TabBarView(
-            children: <Widget>[Text('Todo'), Text('Completed')],
-          ),
-          floatingActionButton: FloatingActionButton(
-            onPressed: () {},
-          ),
-        ),
-      ),
+      home: ToDoScreen(),
     );
-  }
-}
-
-class ToDo extends StatefulWidget {
-  @override
-  _ToDoState createState() => _ToDoState();
-}
-
-class _ToDoState extends State<ToDo> {
-  @override
-  Widget build(BuildContext context) {
-    return Container();
   }
 }
